Refactor issues DONE. Working version 1 DONE.

+++ src/main/kotlin/com/interapp/externa/modules/myAdoptions/infrastructure/RevokeAdoptionPostController.kt
@@ -21,6 +21,7 @@ class RevokeAdoptionPostController(
         val consent = consentService.findConsentByAdoptionId(body.adoptionId)
         consentService.deleteConsent(consent)
         adoptionService.deleteAdoptionById(body.adoptionId)
-        godfatherService.deleteGodfatherById(body.godfatherId)
+        val godfatherAdoptions = adoptionService.findAdoptionsByGodfatherId(body.godfatherId)
+        if (godfatherAdoptions.isEmpty()) godfatherService.deleteGodfatherById(body.godfatherId)
     }
 }
